PostForm Component.
This component contains the form and the post request to generate a new post.

Functions:
handleSubmit - when we submit form, it makes a post request through axios and then injects a innerHTML success message.

handleInputChange - when we change data on a input, it saves it in our state. The state name will be the same as the input name.

NOTE: The post generated will not be saved on the API, but the API returns a response with our data in there.